+++ naga/src/back/wgsl/writer.rs
@@ -1087,25 +1087,21 @@ impl<W: Write> Writer<W> {
         use crate::Expression;
 
         match expressions[expr] {
-            Expression::Literal(literal) => {
-                match literal {
-                    // Floats are written using `Debug` instead of `Display` because it always appends the
-                    // decimal part even it's zero
-                    crate::Literal::F32(value) => write!(self.out, "{:?}", value)?,
-                    crate::Literal::U32(value) => write!(self.out, "{}u", value)?,
-                    crate::Literal::I32(value) => write!(self.out, "{}", value)?,
-                    crate::Literal::Bool(value) => write!(self.out, "{}", value)?,
-                    crate::Literal::F64(value) => write!(self.out, "{:?}lf", value)?,
-                    crate::Literal::I64(_) => {
-                        return Err(Error::Custom("unsupported i64 literal".to_string()));
-                    }
-                    crate::Literal::AbstractInt(_) | crate::Literal::AbstractFloat(_) => {
-                        return Err(Error::Custom(
-                            "Abstract types should not appear in IR presented to backends".into(),
-                        ));
-                    }
+            Expression::Literal(literal) => match literal {
+                crate::Literal::F32(value) => write!(self.out, "{}f", value)?,
+                crate::Literal::U32(value) => write!(self.out, "{}u", value)?,
+                crate::Literal::I32(value) => write!(self.out, "{}i", value)?,
+                crate::Literal::Bool(value) => write!(self.out, "{}", value)?,
+                crate::Literal::F64(value) => write!(self.out, "{:?}lf", value)?,
+                crate::Literal::I64(_) => {
+                    return Err(Error::Custom("unsupported i64 literal".to_string()));
                 }
-            }
+                crate::Literal::AbstractInt(_) | crate::Literal::AbstractFloat(_) => {
+                    return Err(Error::Custom(
+                        "Abstract types should not appear in IR presented to backends".into(),
+                    ));
+                }
+            },
             Expression::Constant(handle) => {
                 let constant = &module.constants[handle];
                 if constant.name.is_some() {